+++ src/docs/examples/Virtualized/index.tsx
@@ -2,15 +2,18 @@ import { Code } from 'docs/components';
 import { Container, Link, Router } from 'uilib';
 
 import * as helpers from './helpers';
+import S from './helpers.styl';
 
 import Endless from '!!raw-loader!./Endless';
 import EndlessScroll from '!!raw-loader!./EndlessScroll';
 import Finite from '!!raw-loader!./Finite';
 import FiniteScroll from '!!raw-loader!./FiniteScroll';
+import InitialTotal from '!!raw-loader!./InitialTotal';
+import ContentBefore from '!!raw-loader!./ContentBefore';
 
 const rootPath = '/virtualized';
 
-const Example = props => <Code scope={{ helpers }} {...props} />;
+const Example = props => <Code scope={{ helpers, S }} {...props} />;
 
 export default () => {
   return (
@@ -28,6 +31,14 @@ export default () => {
             <Link href={`${rootPath}/finite-scroll`}>Finite</Link>
           </div>
         </div>
+        <div>
+          <Link href={`${rootPath}/initial-total`}>
+            Initial data already total
+          </Link>
+        </div>
+        <div>
+          <Link href={`${rootPath}/content-before`}>contentBefore</Link>
+        </div>
       </div>
 
       <Router rootPath={rootPath}>
@@ -44,6 +55,18 @@ export default () => {
           path="/finite-scroll"
           code={FiniteScroll}
           key="finite-scroll"
+        />
+        <Example
+          exact
+          path="/initial-total"
+          code={InitialTotal}
+          key="initial-total"
+        />
+        <Example
+          exact
+          path="/content-before"
+          code={ContentBefore}
+          key="content-before"
         />
       </Router>
     </Container>